+++ packages/backend-core/src/__tests__/apis/OrganizationApi.test.ts
@@ -5,14 +5,18 @@ import { TestBackendAPIClient } from '../TestBackendAPI';
 
 test('createOrganization() creates an organization', async () => {
   const name = 'Acme Inc';
+  const slug = 'acme-inc';
   const createdBy = 'user_randomid';
+  const publicMetadata = { public: 'metadata' };
+  const privateMetadata = { private: 'metadata' };
   const resJSON: OrganizationJSON = {
     object: ObjectType.Organization,
-    public_metadata: {},
     id: 'org_randomid',
     name,
+    slug,
     logo_url: null,
-    slug: null,
+    public_metadata: publicMetadata,
+    private_metadata: privateMetadata,
     created_at: 1611948436,
     updated_at: 1611948436,
   };
@@ -20,13 +24,30 @@ test('createOrganization() creates an organization', async () => {
   nock('https://api.clerk.dev')
     .post('/v1/organizations', {
       name,
+      slug,
+      public_metadata: JSON.stringify(publicMetadata),
+      private_metadata: JSON.stringify(privateMetadata),
       created_by: createdBy,
     })
     .reply(200, resJSON);
 
   const organization = await TestBackendAPIClient.organizations.createOrganization({
     name,
+    slug,
     createdBy,
+    publicMetadata,
+    privateMetadata,
   });
-  expect(organization).toEqual(new Organization(resJSON.id, name, null, null, resJSON.created_at, resJSON.updated_at));
+  expect(organization).toEqual(
+    new Organization(
+      resJSON.id,
+      name,
+      null,
+      null,
+      resJSON.created_at,
+      resJSON.updated_at,
+      publicMetadata,
+      privateMetadata,
+    ),
+  );
 });