+++ packages/backend-core/src/api/utils/ErrorHandler.ts
@@ -6,8 +6,8 @@ export default function handleError(error: any): never {
   const body = error?.response?.body;
   let data;
 
-  if (body && body.errors) {
-    data = (body.errors || []).map((errorJSON: ClerkServerErrorJSON) => {
+  if (body) {
+    data = JSON.parse(body).errors.map((errorJSON: ClerkServerErrorJSON) => {
       return ClerkServerError.fromJSON(errorJSON);
     });
   } else {