+++ packages/clerk-js/src/ui/signIn/utils.ts
@@ -1,5 +1,5 @@
 import { titleize } from '@clerk/shared/utils/string';
-import { PreferredSignInStrategy, SignInFactor, SignInResource } from '@clerk/types';
+import { PreferredSignInStrategy, SignInFactor, SignInResource, SignInStrategy } from '@clerk/types';
 import { PREFERRED_SIGN_IN_STRATEGIES } from 'ui/common';
 import { otpPrefFactorComparator, passwordPrefFactorComparator } from 'ui/signIn/strategies/factorSortingUtils';
 
@@ -100,4 +100,12 @@ export function determineSalutation(signIn: Partial<SignInResource>): string {
   }
 
   return titleize(signIn.userData?.firstName) || titleize(signIn.userData?.lastName) || signIn?.identifier || '';
+}
+
+const localStrategies: SignInStrategy[] = ['email_code', 'password', 'phone_code', 'email_link'];
+export function factorHasLocalStrategy(factor: SignInFactor): boolean {
+  if (!factor) {
+    return false;
+  }
+  return localStrategies.includes(factor.strategy);
 }