+++ packages/remix/src/ssr/rootAuthLoader.ts
@@ -1,10 +1,16 @@
-import { json } from '@remix-run/server-runtime';
-
 import { invalidRootLoaderCallbackResponseReturn, invalidRootLoaderCallbackReturn } from '../errors';
 import { assertFrontendApi } from '../utils';
 import { getAuthData } from './getAuthData';
 import { LoaderFunctionArgs, LoaderFunctionReturn, RootAuthLoaderCallback, RootAuthLoaderOptions } from './types';
-import { assertObject, injectAuthIntoRequest, isRedirect, isResponse, sanitizeAuthData, wrapClerkState } from './utils';
+import {
+  assertObject,
+  injectAuthIntoRequest,
+  isRedirect,
+  isResponse,
+  returnLoaderResultJsonResponse,
+  sanitizeAuthData,
+  throwInterstitialJsonResponse,
+} from './utils';
 
 interface RootAuthLoader {
   <Options extends RootAuthLoaderOptions>(
@@ -30,20 +36,17 @@ export const rootAuthLoader: RootAuthLoader = async (
   const frontendApi = process.env.CLERK_FRONTEND_API || opts.frontendApi;
   assertFrontendApi(frontendApi);
 
-  const { authData, showInterstitial } = await getAuthData(args.request, opts);
+  const { authData, showInterstitial, errorReason } = await getAuthData(args.request, opts);
 
   if (showInterstitial) {
-    throw json(
-      wrapClerkState({ __clerk_ssr_interstitial: showInterstitial, __frontendApi: frontendApi }), 
-      { status: 401 }
-    );
+    throw throwInterstitialJsonResponse({ frontendApi, errorReason });
   }
 
   if (!callback) {
-    return { ...wrapClerkState({ __clerk_ssr_state: authData, __frontendApi: frontendApi }) };
+    return returnLoaderResultJsonResponse({ authData, frontendApi, errorReason });
   }
 
-  const callbackResult = await callback?.(injectAuthIntoRequest(args, sanitizeAuthData(authData!)));
+  const callbackResult = await callback(injectAuthIntoRequest(args, sanitizeAuthData(authData!)));
   assertObject(callbackResult, invalidRootLoaderCallbackReturn);
 
   // Pass through custom responses
@@ -54,5 +57,5 @@ export const rootAuthLoader: RootAuthLoader = async (
     throw new Error(invalidRootLoaderCallbackResponseReturn);
   }
 
-  return { ...callbackResult, ...wrapClerkState({ __clerk_ssr_state: authData, __frontendApi: frontendApi }) };
+  return returnLoaderResultJsonResponse({ authData, frontendApi, errorReason, callbackResult });
 };