+++ packages/edge/CHANGELOG.md
@@ -3,6 +3,38 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+### [1.3.5](https://github.com/clerkinc/javascript/compare/@clerk/edge@1.3.4...@clerk/edge@1.3.5) (2022-04-27)
+
+**Note:** Version bump only for package @clerk/edge
+
+### [1.3.4](https://github.com/clerkinc/javascript/compare/@clerk/edge@1.3.3...@clerk/edge@1.3.4) (2022-04-22)
+
+**Note:** Version bump only for package @clerk/edge
+
+### [1.3.3](https://github.com/clerkinc/javascript/compare/@clerk/edge@1.3.3-staging.1...@clerk/edge@1.3.3) (2022-04-19)
+
+**Note:** Version bump only for package @clerk/edge
+
+### [1.3.3-staging.1](https://github.com/clerkinc/javascript/compare/@clerk/edge@1.3.3-staging.0...@clerk/edge@1.3.3-staging.1) (2022-04-19)
+
+**Note:** Version bump only for package @clerk/edge
+
+### [1.3.2](https://github.com/clerkinc/javascript/compare/@clerk/edge@1.3.2-alpha.0...@clerk/edge@1.3.2) (2022-04-18)
+
+**Note:** Version bump only for package @clerk/edge
+
+### [1.3.2-alpha.0](https://github.com/clerkinc/javascript/compare/@clerk/edge@1.3.1...@clerk/edge@1.3.2-alpha.0) (2022-04-15)
+
+**Note:** Version bump only for package @clerk/edge
+
+### [1.3.1](https://github.com/clerkinc/javascript/compare/@clerk/edge@1.3.0...@clerk/edge@1.3.1) (2022-04-15)
+
+**Note:** Version bump only for package @clerk/edge
+
+## [1.3.0](https://github.com/clerkinc/javascript/compare/@clerk/edge@1.3.0-staging.0...@clerk/edge@1.3.0) (2022-04-15)
+
+**Note:** Version bump only for package @clerk/edge
+
 ### [1.2.7](https://github.com/clerkinc/javascript/compare/@clerk/edge@1.2.7-staging.0...@clerk/edge@1.2.7) (2022-04-13)
 
 **Note:** Version bump only for package @clerk/edge