+++ packages/expo/CHANGELOG.md
@@ -3,6 +3,38 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+### [0.9.15](https://github.com/clerkinc/javascript/compare/@clerk/clerk-expo@0.9.14...@clerk/clerk-expo@0.9.15) (2022-04-27)
+
+**Note:** Version bump only for package @clerk/clerk-expo
+
+### [0.9.14](https://github.com/clerkinc/javascript/compare/@clerk/clerk-expo@0.9.14-staging.0...@clerk/clerk-expo@0.9.14) (2022-04-22)
+
+**Note:** Version bump only for package @clerk/clerk-expo
+
+### [0.9.13](https://github.com/clerkinc/javascript/compare/@clerk/clerk-expo@0.9.13-staging.1...@clerk/clerk-expo@0.9.13) (2022-04-19)
+
+**Note:** Version bump only for package @clerk/clerk-expo
+
+### [0.9.13-staging.1](https://github.com/clerkinc/javascript/compare/@clerk/clerk-expo@0.9.13-staging.0...@clerk/clerk-expo@0.9.13-staging.1) (2022-04-19)
+
+**Note:** Version bump only for package @clerk/clerk-expo
+
+### [0.9.12](https://github.com/clerkinc/javascript/compare/@clerk/clerk-expo@0.9.12-alpha.0...@clerk/clerk-expo@0.9.12) (2022-04-18)
+
+**Note:** Version bump only for package @clerk/clerk-expo
+
+### [0.9.12-alpha.0](https://github.com/clerkinc/javascript/compare/@clerk/clerk-expo@0.9.11...@clerk/clerk-expo@0.9.12-alpha.0) (2022-04-15)
+
+**Note:** Version bump only for package @clerk/clerk-expo
+
+### [0.9.11](https://github.com/clerkinc/javascript/compare/@clerk/clerk-expo@0.9.10...@clerk/clerk-expo@0.9.11) (2022-04-15)
+
+**Note:** Version bump only for package @clerk/clerk-expo
+
+### [0.9.10](https://github.com/clerkinc/javascript/compare/@clerk/clerk-expo@0.9.10-staging.1...@clerk/clerk-expo@0.9.10) (2022-04-15)
+
+**Note:** Version bump only for package @clerk/clerk-expo
+
 ### [0.9.9](https://github.com/clerkinc/javascript/compare/@clerk/clerk-expo@0.9.9-staging.0...@clerk/clerk-expo@0.9.9) (2022-04-13)
 
 **Note:** Version bump only for package @clerk/clerk-expo