+++ packages/nextjs/CHANGELOG.md
@@ -3,6 +3,42 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [3.4.0](https://github.com/clerkinc/javascript/compare/@clerk/nextjs@3.3.7...@clerk/nextjs@3.4.0) (2022-04-27)
+
+### Features
+
+- **nextjs:** Fix opts parsing in withServerSideAuth ([1eecb81](https://github.com/clerkinc/javascript/commit/1eecb81439809fd4536d4b7d2f65d4fed6c6137b))
+
+### [3.3.7](https://github.com/clerkinc/javascript/compare/@clerk/nextjs@3.3.6...@clerk/nextjs@3.3.7) (2022-04-22)
+
+### Bug Fixes
+
+- **nextjs,backend-core:** Fix handleError parsing ([4eb9732](https://github.com/clerkinc/javascript/commit/4eb97324d9153aa91492982d00bb76b8592947c7))
+
+### [3.3.6](https://github.com/clerkinc/javascript/compare/@clerk/nextjs@3.3.6-staging.1...@clerk/nextjs@3.3.6) (2022-04-19)
+
+**Note:** Version bump only for package @clerk/nextjs
+
+### [3.3.6-staging.1](https://github.com/clerkinc/javascript/compare/@clerk/nextjs@3.3.6-staging.0...@clerk/nextjs@3.3.6-staging.1) (2022-04-19)
+
+**Note:** Version bump only for package @clerk/nextjs
+
+### [3.3.5](https://github.com/clerkinc/javascript/compare/@clerk/nextjs@3.3.5-alpha.0...@clerk/nextjs@3.3.5) (2022-04-18)
+
+**Note:** Version bump only for package @clerk/nextjs
+
+### [3.3.5-alpha.0](https://github.com/clerkinc/javascript/compare/@clerk/nextjs@3.3.4...@clerk/nextjs@3.3.5-alpha.0) (2022-04-15)
+
+**Note:** Version bump only for package @clerk/nextjs
+
+### [3.3.4](https://github.com/clerkinc/javascript/compare/@clerk/nextjs@3.3.3...@clerk/nextjs@3.3.4) (2022-04-15)
+
+**Note:** Version bump only for package @clerk/nextjs
+
+### [3.3.3](https://github.com/clerkinc/javascript/compare/@clerk/nextjs@3.3.3-staging.1...@clerk/nextjs@3.3.3) (2022-04-15)
+
+**Note:** Version bump only for package @clerk/nextjs
+
 ### [3.3.2](https://github.com/clerkinc/javascript/compare/@clerk/nextjs@3.3.2-staging.0...@clerk/nextjs@3.3.2) (2022-04-13)
 
 **Note:** Version bump only for package @clerk/nextjs