+++ packages/shared/CHANGELOG.md
@@ -3,6 +3,26 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+### [0.1.4](https://github.com/clerkinc/clerk_docker/compare/@clerk/shared@0.1.4-staging.1...@clerk/shared@0.1.4) (2022-04-19)
+
+**Note:** Version bump only for package @clerk/shared
+
+### [0.1.4-staging.1](https://github.com/clerkinc/clerk_docker/compare/@clerk/shared@0.1.4-staging.0...@clerk/shared@0.1.4-staging.1) (2022-04-19)
+
+**Note:** Version bump only for package @clerk/shared
+
+### [0.1.3](https://github.com/clerkinc/clerk_docker/compare/@clerk/shared@0.1.3-alpha.0...@clerk/shared@0.1.3) (2022-04-18)
+
+**Note:** Version bump only for package @clerk/shared
+
+### [0.1.3-alpha.0](https://github.com/clerkinc/clerk_docker/compare/@clerk/shared@0.1.2...@clerk/shared@0.1.3-alpha.0) (2022-04-15)
+
+**Note:** Version bump only for package @clerk/shared
+
+### [0.1.2](https://github.com/clerkinc/clerk_docker/compare/@clerk/shared@0.1.2-staging.0...@clerk/shared@0.1.2) (2022-04-15)
+
+**Note:** Version bump only for package @clerk/shared
+
 ### [0.1.1](https://github.com/clerkinc/clerk_docker/compare/@clerk/shared@0.1.1-staging.0...@clerk/shared@0.1.1) (2022-04-13)
 
 **Note:** Version bump only for package @clerk/shared